Implemented functionality to perform the operations based on logged in user

+++ todo-app/src/main/java/com/sps/todoapp/controller/TodoController.java
@@ -1,14 +1,12 @@
 package com.sps.todoapp.controller;
 
-import java.util.List;
-
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import com.sps.todoapp.exception.ResourceNotFoundException;
 import com.sps.todoapp.model.Todo;
+import com.sps.todoapp.model.User;
 import com.sps.todoapp.service.TodoService;
 
 @Controller
@@ -28,14 +27,25 @@ public class TodoController {
 
 	@GetMapping("/todoList")
 	public String getAllTodos(ModelMap model, HttpServletRequest req) {
-		req.setAttribute("todos", service.getAllTodos());
+		HttpSession session = req.getSession(false);
+		if (null == session || null == session.getAttribute("user")) {
+			return "login";
+		}
+		User user = (User) session.getAttribute("user");
+		req.setAttribute("todos", service.getAllTodos(user.getId()));
 		return "todo/todoList";
 	}
 
 	@PostMapping("/todoAdd")
 	public String createTodo(@ModelAttribute Todo todo, BindingResult result, HttpServletRequest req) {
+		HttpSession session = req.getSession(false);
+		if (null == session || null == session.getAttribute("user")) {
+			return "login";
+		}
+		User user = (User) session.getAttribute("user");
+		todo.setUserId(user.getId());
 		service.create(todo);
-		req.setAttribute("todos", service.getAllTodos());
+		req.setAttribute("todos", service.getAllTodos(user.getId()));
 		req.setAttribute("successMsg", "Task added successfully");
 		return "todo/todoList";
 	}
@@ -50,8 +60,13 @@ public class TodoController {
 
 	@PostMapping("/todoUpdate")
 	public String updateTodo(@ModelAttribute Todo todo, BindingResult result, HttpServletRequest req) {
+		HttpSession session = req.getSession(false);
+		if (null == session || null == session.getAttribute("user")) {
+			return "login";
+		}
+		User user = (User) session.getAttribute("user");
 		service.update(todo, null);
-		req.setAttribute("todos", service.getAllTodos());
+		req.setAttribute("todos", service.getAllTodos(user.getId()));
 		req.setAttribute("successMsg", "Task updated successfully");
 		return "todo/todoList";
 	}
@@ -64,6 +79,7 @@ public class TodoController {
 		todo.setDescription(dbTodo.getDescription());
 		todo.setStatus(dbTodo.getStatus());
 		todo.setTargetDate(dbTodo.getTargetDate());
+		todo.setUserId(dbTodo.getUserId());
 		req.setAttribute("todo", dbTodo);
 		return "todo/todoAdd";
 
@@ -71,8 +87,13 @@ public class TodoController {
 
 	@GetMapping("/todoDelete")
 	public String deleteTodo(@RequestParam("id") Long id, HttpServletRequest req) throws ResourceNotFoundException {
+		HttpSession session = req.getSession(false);
+		if (null == session || null == session.getAttribute("user")) {
+			return "login";
+		}
+		User user = (User) session.getAttribute("user");
 		service.delete(id);
-		req.setAttribute("todos", service.getAllTodos());
+		req.setAttribute("todos", service.getAllTodos(user.getId()));
 		req.setAttribute("successMsg", "Task deleted successfully");
 		return "todo/todoList";
 	}
